pyproject.toml
[project]
name = "financial-agent-crew"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai>=1.9.3",
    "crewai-tools>=1.9.3",
    "fastapi>=0.128.6",
    "langchain-google-genai>=4.2.0",
    "pandas-ta>=0.4.71b0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.40.0",
    "yfinance>=1.1.0",
]
